refactor: extract Header, Editor, Footer components
