+++ src/components/Header.tsx
@@ -1,0 +1,5 @@
+export default function Header() {
+  return (
+    <h1 className="h1"><span className="badge">⚖️</span> Qaadi Live</h1>
+  );
+}
